requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
environs==14.3.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
marshmallow==4.0.1
multidict==6.6.4
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
typing-inspection==0.4.1
typing_extensions==4.15.0
yarl==1.20.1
